Zolfekar3

+++ anime-flutter/anime_galaxy/lib/module_episode/request/rating_request/rating_request.dart
@@ -1,7 +1,7 @@
 class RatingRequest{
   String userId;
   int episodeId;
-  double rateValue;
+  int rateValue;
 
   RatingRequest({
     this.userId,
